test: AdminController & CafeController CustomGiven 적용
